vim <--> clavm <--> tcc; compiling is working

+++ SharedMemoryMacOS.cpp
@@ -1,9 +1,10 @@
 #include <fcntl.h>
 #include <semaphore.h>
-#include <sys/mman.h>
-#include <unistd.h>
 #include <stdlib.h>
 #include <string.h>
+#include <sys/mman.h>
+#include <unistd.h>
+
 #include <iostream>
 
 class _SharedMemory;
@@ -11,13 +12,15 @@ class _SharedMemory {
   void* _shared;
   size_t _size;
   const char* _name;
+  bool _create;
   int _file;
 
  public:
-  _SharedMemory(const char* name, size_t size, bool create) : _shared(nullptr), _size(size), _name(name) {
+  _SharedMemory(const char* name, size_t size, bool create)
+      : _shared(nullptr), _size(size), _name(name), _create(create) {
     // XXX possibly unlink here
 
-    if (create) {
+    if (_create) {
       _file = shm_open(name, O_CREAT | O_RDWR, 0666);
     } else {
       _file = shm_open(name, O_RDWR, 0666);
@@ -30,7 +33,7 @@ class _SharedMemory {
 
     // XXX just on create?
     //
-    if (create) {
+    if (_create) {
       // set size and fill with zeros
       ftruncate(_file, _size);
     }
@@ -43,7 +46,9 @@ class _SharedMemory {
   }
 
   ~_SharedMemory() {
-    unlink(_name);
+    if (_create) {
+      unlink(_name);
+    }
     munmap(_shared, _size);
     close(_file);
   }
